+++ include/fi_util.h
@@ -54,7 +54,7 @@
 
 #include <fi.h>
 #include <ofi_mr.h>
-#include <fi_list.h>
+#include <ofi_list.h>
 #include <ofi_mem.h>
 #include <ofi_rbuf.h>
 #include <ofi_signal.h>
