+++ include/ofi_mr.h
@@ -43,7 +43,7 @@
 #include <fi.h>
 #include <ofi_atom.h>
 #include <ofi_lock.h>
-#include <fi_list.h>
+#include <ofi_list.h>
 #include <rbtree.h>
