+++ libfabric.vcxproj
@@ -574,7 +574,7 @@
     <ClInclude Include="include\ofi_file.h" />
     <ClInclude Include="include\ofi_iov.h" />
     <ClInclude Include="include\ofi_indexer.h" />
-    <ClInclude Include="include\fi_list.h" />
+    <ClInclude Include="include\ofi_list.h" />
     <ClInclude Include="include\ofi_lock.h" />
     <ClInclude Include="include\ofi_mem.h" />
     <ClInclude Include="include\fi_osd.h" />
