+++ prov/gni/include/gnix_cq.h
@@ -41,7 +41,7 @@
 #include "gnix_queue.h"
 #include "gnix_wait.h"
 #include "gnix_util.h"
-#include <fi_list.h>
+#include <ofi_list.h>
 #include <stdbool.h>
 
 #define GNIX_CQ_DEFAULT_FORMAT struct fi_cq_entry
