+++ prov/gni/include/gnix_freelist.h
@@ -35,7 +35,7 @@
 #define _GNIX_FREELIST_H_
 
 #include <fi.h>
-#include <fi_list.h>
+#include <ofi_list.h>
 #include "include/gnix_util.h"
 
 /* Number of elements to seed the freelist with */
