+++ prov/gni/include/gnix_smrn.h
@@ -33,7 +33,7 @@
 #ifndef PROV_GNI_INCLUDE_GNIX_SMRN_H
 #define PROV_GNI_INCLUDE_GNIX_SMRN_H
 
-#include "include/fi_list.h"
+#include "include/ofi_list.h"
 #include "include/ofi_lock.h"
 #include "config.h"
