+++ prov/gni/include/gnix_tags.h
@@ -151,7 +151,7 @@
 
 #include <stdlib.h>
 #include <fi.h>
-#include <fi_list.h>
+#include <ofi_list.h>
 
 #include "gnix_util.h"
