+++ prov/psm/src/psmx.h
@@ -66,7 +66,7 @@ extern "C" {
 #include "fi.h"
 #include "ofi_atomic.h"
 #include "fi_enosys.h"
-#include "fi_list.h"
+#include "ofi_list.h"
 #include "fi_util.h"
 #include "ofi_file.h"
 #include "rbtree.h"
