+++ prov/psm2/src/psmx2.h
@@ -67,7 +67,7 @@ extern "C" {
 #include "fi.h"
 #include "ofi_atomic.h"
 #include "fi_enosys.h"
-#include "fi_list.h"
+#include "ofi_list.h"
 #include "fi_util.h"
 #include "ofi_mem.h"
 #include "rbtree.h"
