+++ prov/shm/src/smr.h
@@ -55,7 +55,7 @@
 #include <fi_enosys.h>
 #include <ofi_shm.h>
 #include <ofi_rbuf.h>
-#include <fi_list.h>
+#include <ofi_list.h>
 #include <ofi_signal.h>
 #include <fi_util.h>
 #include <ofi_atomic.h>
