+++ prov/udp/src/udpx.h
@@ -56,7 +56,7 @@
 #include <fi.h>
 #include <fi_enosys.h>
 #include <ofi_rbuf.h>
-#include <fi_list.h>
+#include <ofi_list.h>
 #include <ofi_signal.h>
 #include <fi_util.h>
