+++ prov/verbs/src/ep_rdm/verbs_cq_ep_rdm.c
@@ -33,7 +33,7 @@
 #include <stdlib.h>
 
 #include <fi_enosys.h>
-#include <fi_list.h>
+#include <ofi_list.h>
 #include "../fi_verbs.h"
 #include "verbs_queuing.h"
