+++ prov/verbs/src/ep_rdm/verbs_ep_rdm.c
@@ -35,7 +35,7 @@
 #include <arpa/inet.h>
 #include <rdma/rdma_cma.h>
 
-#include <fi_list.h>
+#include <ofi_list.h>
 #include <ofi_file.h>
 #include <fi_enosys.h>
 #include "../fi_verbs.h"
