+++ prov/verbs/src/ep_rdm/verbs_tagged_ep_rdm.c
@@ -34,7 +34,7 @@
 
 #include <config.h>
 
-#include <fi_list.h>
+#include <ofi_list.h>
 #include <fi_enosys.h>
 #include <ofi_iov.h>
 #include <rdma/fi_tagged.h>
