+++ src/var.c
@@ -40,7 +40,7 @@
 #include <rdma/fi_errno.h>
 
 #include "fi.h"
-#include "fi_list.h"
+#include "ofi_list.h"
 
 
 extern int ofi_init;
